src: gpu: conv: enable f64 float immediate expr

+++ src/gpu/jit/conv/ir_simplify.cpp
@@ -2168,7 +2168,7 @@ expr_t const_fold_non_recursive(const expr_t &e) {
         if (cast->expr.is<int_imm_t>())
             return to_expr(to_cpp<int64_t>(cast->expr), cast->type);
         if (cast->expr.is<float_imm_t>())
-            return to_expr(to_cpp<float>(cast->expr), cast->type);
+            return to_expr(to_cpp<double>(cast->expr), cast->type);
     }
 
     return e;
